requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyautogui>=0.9.54
psutil>=5.9.0
opencv-python>=4.8.0
pytesseract>=0.3.10
wikipedia>=1.4.0
requests>=2.31.0
screen-brightness-control>=0.23.0
Pillow>=10.0.0
pycaw>=20240210
comtypes
